Rename _io.py to _input_output.py
There is a built-in Python module called `_io.py` and our internal
private module was clashing with that name. Rename it to
`_input_output.py`. The longer name is not a problem since it's only
used internally and not in the API.
