+++ magali/_input_output.py
